pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kadbot"
version = "1.0.0"
description = "Парсер дел с сайта kad.arbitr.ru с интеграцией в Aspro CRM"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "undetected-chromedriver>=3.5.5",
    "selenium>=4.21.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "SQLAlchemy>=2.0.23",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.1",
    "flake8>=7.0.0",
    "isort>=5.13.2",
    "pylint>=3.0.3",
    "mypy>=1.7.1",
    "types-requests>=2.31.0.20231231",
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=2.0.0",
    "pre-commit>=3.6.0",
    "jupyter>=1.0.0",
    "ipython>=8.18.1",
]

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 79
known_first_party = ["kadbot"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=kadbot",
    "--cov-report=term-missing",
    "--cov-report=html",
]
